pyproject.toml
[tool.poetry]
name = "django23"
version = "0.1.0"
description = ""
authors = ["Alexandr <blix@list.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
pandas = "^2.2.3"
openpyxl = "^3.1.5"
flake8 = "^7.1.1"
black = "^24.10.0"
pytest = "^8.3.3"
pytest-cov = "^5.0.0"
mypy = "^1.13.0"
psycopg = "^3.2.3"
psycopg2 = "^2.9.10"
isort = "^5.13.2"
config = "^0.5.1"
iniconfig = "^2.0.0"
django = "^5.1.2"
ipython = "^8.28.0"
pillow = "^11.0.0"
redis = "^5.2.0"
djangorestframework = "^3.15.2"
django-filter = "^24.3"
djangorestframework-simplejwt = "^5.4.0"
coverage = "^7.6.10"
validators = "^0.34.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
